Finish extracting univariate_distributions into separate component.
